requirements.txt
# =============================================================================
# Dépendances - Agent d'analyse de marché e-commerce
# =============================================================================

# --- Web Framework ---
fastapi==0.135.1              # API REST (async, validation, Swagger auto-généré)
uvicorn==0.42.0               # Serveur ASGI pour FastAPI

# --- LangGraph & LangChain ---
langgraph==0.2.60             # Orchestration de l'agent (graphe)
langchain-core==0.3.83        # Abstractions: messages, outils, prompts
langchain-groq==0.3.8        # Client LLM Groq (gratuit)

# --- Data Validation ---
pydantic==2.12.5              # Validation des données (utilisé par FastAPI)

# --- Utilities ---
python-dotenv==1.0.1          # Charger les variables d'environnement

# --- Testing ---
pytest==8.3.3                 # Framework de tests
httpx==0.27.2                 # Client HTTP async (requis par FastAPI TestClient)
